uint flags
