+++ ask.go
@@ -2,7 +2,6 @@ package ask
 
 import (
 	"context"
-	"encoding/hex"
 	"errors"
 	"flag"
 	"fmt"
@@ -755,6 +754,10 @@ func FlagValue(typ reflect.Type, val reflect.Value) (flag.Value, error) {
 				case reflect.Uint8:
 					b := (*[]byte)(ptr)
 					fl = (*BytesHexFlag)(b)
+				case reflect.Uint16:
+					fl = (*Uint16SliceValue)(ptr)
+				case reflect.Uint32:
+					fl = (*Uint32SliceValue)(ptr)
 				case reflect.Uint:
 					fl = (*UintSliceValue)(ptr)
 				case reflect.Int:
@@ -803,121 +806,4 @@ func FlagValue(typ reflect.Type, val reflect.Value) (flag.Value, error) {
 		}
 	}
 	return fl, nil
-}
-
-// BytesHex exposes bytes as a flag, hex-encoded,
-// optional whitespace padding, case insensitive, and optional 0x prefix.
-type BytesHexFlag []byte
-
-func (f BytesHexFlag) String() string {
-	return hex.EncodeToString(f)
-}
-
-func (f *BytesHexFlag) Set(value string) error {
-	value = strings.TrimSpace(value)
-	value = strings.ToLower(value)
-	if strings.HasPrefix(value, "0x") {
-		value = value[2:]
-	}
-	b, err := hex.DecodeString(value)
-	if err != nil {
-		return err
-	}
-	*f = b
-	return nil
-}
-
-func (f *BytesHexFlag) Type() string {
-	return "bytes"
-}
-
-// fixedLenBytes exposes fixed-length bytes as a flag, hex-encoded,
-// optional whitespace padding, case insensitive, and optional 0x prefix.
-type fixedLenBytes struct {
-	Dest           []byte
-	ExpectedLength uint64
-}
-
-func (f fixedLenBytes) String() string {
-	return hex.EncodeToString(f.Dest)
-}
-
-func (f *fixedLenBytes) Set(value string) error {
-	value = strings.TrimSpace(value)
-	value = strings.ToLower(value)
-	if strings.HasPrefix(value, "0x") {
-		value = value[2:]
-	}
-	b, err := hex.DecodeString(value)
-	if err != nil {
-		return err
-	}
-	if uint64(len(b)) != f.ExpectedLength {
-		return fmt.Errorf("byte length does not match fixed-length of %d bytes: "+
-			"parsed %d bytes", f.ExpectedLength, len(b))
-	}
-	copy(f.Dest, b)
-	return nil
-}
-
-func (f *fixedLenBytes) Type() string {
-	return fmt.Sprintf("bytes%d", f.ExpectedLength)
-}
-
-// fixedLenBytesSlice exposes a slice of fixed-length bytes elements as a flag,
-// optional whitespace/padding, comma-separated.
-// Each element is hex-encoded, case insensitive, and optional 0x prefix.
-type fixedLenBytesSlice struct {
-	Dest reflect.Value
-}
-
-func (f fixedLenBytesSlice) String() string {
-	var buf strings.Builder
-	length := f.Dest.Len()
-	elemLen := f.Dest.Type().Elem().Len()
-	for i := 0; i < length; i++ {
-		if i > 0 {
-			buf.WriteString(",")
-		}
-		item := f.Dest.Index(i)
-		itemBytes := item.Slice(0, elemLen).Bytes()
-		buf.WriteString(hex.EncodeToString(itemBytes))
-	}
-	return buf.String()
-}
-
-func (f *fixedLenBytesSlice) Set(value string) error {
-	value = strings.TrimSpace(value)
-	value = strings.ToLower(value)
-	var elems []string
-	if len(value) > 0 {
-		elems = strings.Split(value, ",")
-	}
-	dest := reflect.MakeSlice(f.Dest.Type(), len(elems), len(elems))
-	elemTyp := f.Dest.Type().Elem()
-	elemLen := elemTyp.Len()
-	for i, el := range elems {
-		el = strings.TrimSpace(el)
-		if strings.HasPrefix(el, "0x") {
-			el = el[2:]
-		}
-		b, err := hex.DecodeString(el)
-		if err != nil {
-			return err
-		}
-		if len(b) != elemLen {
-			return fmt.Errorf("byte length of element %d does not match fixed-length of %d bytes: "+
-				"parsed %d bytes", i, elemLen, len(b))
-		}
-		destElem := dest.Index(i)
-		destElemBytes := destElem.Slice(0, elemLen).Bytes()
-		copy(destElemBytes, b)
-	}
-	f.Dest.Set(dest)
-	return nil
-}
-
-func (f *fixedLenBytesSlice) Type() string {
-	elemLen := f.Dest.Type().Elem().Len()
-	return fmt.Sprintf("[]bytes%d", elemLen)
 }
